# Conflicts:
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/endpoint/MessageEndpoint.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/repository/GroupRepository.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/repository/UserGroupRepository.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/repository/UserRepository.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/UserGroupService.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/UserService.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/CustomUserDetailService.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/GroupServiceImpl.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/SimpleMessageService.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/UserGroupServiceImpl.java
#	backend/src/test/java/at/ac/tuwien/sepr/groupphase/backend/endpoint/test/GroupEndpointTest.java
#	backend/src/test/java/at/ac/tuwien/sepr/groupphase/backend/repository/test/UserGroupRepositoryTest.java
#	backend/src/test/java/at/ac/tuwien/sepr/groupphase/backend/service/test/GroupServiceImplTest.java
#	frontend/src/app/app.module.ts
#	frontend/src/app/components/groups/group-create-edit/group-create-edit.component.html
#	frontend/src/app/components/groups/group-create-edit/group-create-edit.component.ts
#	frontend/src/app/components/groups/group-detail/group-detail.component.ts
#	frontend/src/app/components/login/login.component.ts
#	frontend/src/app/components/message/message.component.html
#	frontend/src/app/components/message/message.component.ts
#	frontend/src/app/confirmation-dialog/confirmation-dialog.component.spec.ts
#	frontend/src/app/interceptors/auth-interceptor.ts
#	frontend/src/app/services/dialog.service.ts
#	frontend/src/app/services/groups.service.ts
#	frontend/src/app/services/user.service.ts

+++ backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/SimpleMessageService.java
@@ -3,21 +3,24 @@ package at.ac.tuwien.sepr.groupphase.backend.service.impl;
 import at.ac.tuwien.sepr.groupphase.backend.endpoint.dto.MessageCreateDto;
 import at.ac.tuwien.sepr.groupphase.backend.endpoint.dto.MessageDetailDto;
 import at.ac.tuwien.sepr.groupphase.backend.entity.ApplicationMessage;
+import at.ac.tuwien.sepr.groupphase.backend.entity.ApplicationUser;
 import at.ac.tuwien.sepr.groupphase.backend.exception.ConflictException;
 import at.ac.tuwien.sepr.groupphase.backend.exception.NotFoundException;
 import at.ac.tuwien.sepr.groupphase.backend.exception.ValidationException;
-import at.ac.tuwien.sepr.groupphase.backend.repository.GroupRepository;
 import at.ac.tuwien.sepr.groupphase.backend.repository.MessageRepository;
-import at.ac.tuwien.sepr.groupphase.backend.repository.UserRepository;
 import at.ac.tuwien.sepr.groupphase.backend.service.MessageService;
 
 import java.lang.invoke.MethodHandles;
 import java.time.LocalDateTime;
 import java.util.List;
 
+import at.ac.tuwien.sepr.groupphase.backend.service.UserService;
+import at.ac.tuwien.sepr.groupphase.backend.service.validators.MessageValidator;
+import org.hibernate.exception.ConstraintViolationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 /**
@@ -29,25 +32,29 @@ public class SimpleMessageService implements MessageService {
     private static final Logger LOGGER =
         LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
     private final MessageRepository messageRepository;
+    private final UserService userService;
+    private final MessageValidator validator;
 
     @Autowired
-    private UserRepository userRepository;
-
-    public SimpleMessageService(MessageRepository messageRepository) {
+    public SimpleMessageService(MessageRepository messageRepository, UserService userService, MessageValidator validator) {
         this.messageRepository = messageRepository;
+        this.userService = userService;
+        this.validator = validator;
     }
 
     @Override
-    public List<ApplicationMessage> findAll() {
+    public List<ApplicationMessage> findAll() throws NotFoundException {
         LOGGER.debug("Find all messages");
-        return messageRepository.findAllByOrderByIsReadAscSentAtDesc();
+        ApplicationUser user = userService.findApplicationUserByEmail(SecurityContextHolder.getContext().getAuthentication().getName());
+        return messageRepository.findAllByApplicationUserOrderByIsReadAscSentAtDesc(user);
     }
 
     @Override
-    public ApplicationMessage create(MessageCreateDto message) {
+    public ApplicationMessage create(MessageCreateDto message) throws ConstraintViolationException, ValidationException {
         LOGGER.debug("Publish new message {}", message);
+        validator.validateForCreate(message);
         ApplicationMessage applicationMessage = ApplicationMessage.ApplicationMessageBuilder.message()
-            .withApplicationUser(userRepository.findByEmail("user1@email.com"))
+            .withApplicationUser(userService.findApplicationUserById(message.getUserId()))
             .withGroupId(message.getGroupId())
             .withIsRead(false)
             .withSentAt(LocalDateTime.now())
@@ -58,14 +65,24 @@ public class SimpleMessageService implements MessageService {
     @Override
     public ApplicationMessage update(MessageDetailDto toUpdate) throws NotFoundException, ValidationException, ConflictException {
         LOGGER.debug("Update message {}", toUpdate);
-        ApplicationMessage myMessage = messageRepository.findById(toUpdate.getId()).orElseThrow();
-        myMessage.setIsRead(true);
+        ApplicationMessage myMessage = messageRepository.findById(toUpdate.getId()).orElse(null);
+        if (myMessage == null) {
+            throw new NotFoundException(String.format("Could not find message from group %s sent at %s",
+                toUpdate.getGroup().getName(), toUpdate.getSentAt()));
+        }
+        toUpdate.setIsRead(true);
+        validator.validateForUpdate(toUpdate);
+        myMessage.setIsRead(toUpdate.getIsRead());
         return messageRepository.save(myMessage);
     }
 
     @Override
-    public void delete(Long messageId) {
+    public void delete(Long messageId) throws NotFoundException {
         LOGGER.debug("Delete message with id {}", messageId);
+        if (!messageRepository.existsById(messageId)) {
+            throw new NotFoundException(String.format("Could not find message with id %s",
+                messageId));
+        }
         messageRepository.deleteById(messageId);
     }
